requirements.txt
Django>=4.1.7,<5.0
djangorestframework>=3.14.0,<4.0
requests>=2.28.2,<3.0
mysql-connector-python>=8.0.32,<9.0
mysqlclient>=2.1.1,<3.0
python-dotenv
